extract MyStatePhoneListener
