+++ app/src/main/java/org/stepic/droid/core/StepicCoreComponent.java
@@ -34,7 +34,6 @@ import org.stepic.droid.view.dialogs.LogoutAreYouSureDialog;
 import org.stepic.droid.view.dialogs.RemindPasswordDialogFragment;
 import org.stepic.droid.view.dialogs.VideoQualityDialog;
 import org.stepic.droid.view.fragments.DownloadsFragment;
-import org.stepic.droid.view.fragments.VideoFragment;
 import org.stepic.droid.web.RetrofitRESTApi;
 
 import javax.inject.Singleton;
@@ -119,5 +118,5 @@ public interface StepicCoreComponent {
 
     void inject(FromDbCoursesTask fromDbCoursesTask);
 
-    void inject(VideoFragment.MyStatePhoneListener receiver);
+    void inject(MyStatePhoneListener listener);
 }
