+++ app/src/main/java/org/stepic/droid/view/fragments/VideoFragment.kt
@@ -16,13 +16,12 @@ import android.telephony.TelephonyManager
 import android.util.Log
 import android.view.*
 import android.widget.*
-import com.squareup.otto.Bus
 import com.squareup.otto.Subscribe
 import com.yandex.metrica.YandexMetrica
 import org.stepic.droid.R
 import org.stepic.droid.base.FragmentBase
 import org.stepic.droid.base.MainApplication
-import org.stepic.droid.concurrency.IMainHandler
+import org.stepic.droid.core.MyStatePhoneListener
 import org.stepic.droid.events.IncomingCallEvent
 import org.stepic.droid.events.audio.AudioFocusLossEvent
 import org.stepic.droid.preferences.VideoPlaybackRate
@@ -36,7 +35,6 @@ import org.videolan.libvlc.MediaPlayer
 import org.videolan.libvlc.util.AndroidUtil
 import java.io.File
 import java.util.*
-import javax.inject.Inject
 
 class VideoFragment : FragmentBase(), LibVLC.HardwareAccelerationError, IVLCVout.Callback {
     companion object {
@@ -960,24 +958,7 @@ class VideoFragment : FragmentBase(), LibVLC.HardwareAccelerationError, IVLCVout
         }
     }
 
-    class MyStatePhoneListener : PhoneStateListener() {
 
-        init {
-            MainApplication.component().inject(this)
-        }
-
-        @Inject
-        lateinit var mBus: Bus
-
-        @Inject
-        lateinit var mHandler: IMainHandler
-
-        override fun onCallStateChanged(state: Int, incomingNumber: String?) {
-            if (state == 1) {
-                mHandler.post { mBus.post(IncomingCallEvent()) }
-            }
-        }
-    }
 
     @Subscribe
     fun onAudioFocusLoss(event: AudioFocusLossEvent) {
